move to _pixelmap
