+++ py/circuitpy_defns.mk
@@ -267,7 +267,7 @@ ifeq ($(CIRCUITPY_PIXELBUF),1)
 SRC_PATTERNS += adafruit_pixelbuf/%
 endif
 ifeq ($(CIRCUITPY_PIXELMAP),1)
-SRC_PATTERNS += adafruit_pixelmap/%
+SRC_PATTERNS += _pixelmap/%
 endif
 ifeq ($(CIRCUITPY_QRIO),1)
 SRC_PATTERNS += qrio/%
@@ -546,8 +546,8 @@ SRC_SHARED_MODULE_ALL = \
 	_eve/__init__.c \
 	adafruit_pixelbuf/PixelBuf.c \
 	adafruit_pixelbuf/__init__.c \
-	adafruit_pixelmap/PixelMap.c \
-	adafruit_pixelmap/__init__.c \
+	_pixelmap/PixelMap.c \
+	_pixelmap/__init__.c \
 	_stage/Layer.c \
 	_stage/Text.c \
 	_stage/__init__.c \
