+++ shared-bindings/_pixelmap/PixelMap.h
