+++ shared-bindings/_pixelmap/__init__.h
