+++ shared-module/_pixelmap/PixelMap.c
@@ -28,9 +28,9 @@
 #include "py/smallint.h"
 #include "py/runtime.h"
 
-#include "shared-bindings/adafruit_pixelmap/PixelMap.h"
+#include "shared-bindings/_pixelmap/PixelMap.h"
 #include "shared-bindings/adafruit_pixelbuf/PixelBuf.h"
-#include "shared-module/adafruit_pixelmap/PixelMap.h"
+#include "shared-module/_pixelmap/PixelMap.h"
 
 
 static void pixelmap_set_pixel_rgbw(pixelmap_pixelmap_obj_t *self, size_t i, color_u rgbw) {
