+++ shared-module/_pixelmap/PixelMap.h
